+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ### v0.1.4
 
 - Add CdnClusterGet and CndNodeGet methods
+- Add Address and PrivateKeyHex method to crypto